<!--- Provide a general summary of your changes in the Title above -->

## Description
This PR removes some of the odd choices built into the RDF class.

## Motivation and Context
The use of off-center "bin centers" is not well motivated and not worth keeping. Neither is the automatic zeroing out of the initial bin of the RDF, or the first two cumulative bins. There is some question of the convention used for the cumulative count (whether or not the zeroth bin should be empty), but since we're taking discrete sums and not integrating it makes much more sense to have a nonempty zero bin and have the final bin actually be the total.

This PR is still a draft because there is one other change I want to investigate before finalizing. I also need to update the documentation for the cumulative counts.

## How Has This Been Tested?
Existing tests were tailored to the current output, which in hindsight is probably unexpected. I've updated tests to account for the new behaviors.

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds or improves functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [**CONTRIBUTING**](https://github.com/glotzerlab/freud/blob/master/CONTRIBUTING.md) document.
- [x] My code follows the code style of this project.
- [ ] I have updated the documentation (if relevant).
- [x] I have added tests that cover my changes (if relevant).
- [x] All new and existing tests passed.
- [ ] I have updated the [Changelog](https://github.com/glotzerlab/freud/blob/master/ChangeLog.md).