@vyasr 2019-09-04
I ran the benchmarks for the RDF, except I set `dr=0.001` to get 10000 bins. Here are the results of that benchmark both with and without the `parallel_scan` for accumulation:

With `parallel_scan`:
```
Threads          1000          |         10000          |
      1   21.142 ms      1.00x | 1181.440 ms      1.00x |
      2    9.304 ms      2.27x |  636.024 ms      1.86x |
      3    6.910 ms      3.06x |  421.282 ms      2.80x |
      4    5.892 ms      3.59x |  345.159 ms      3.42x |
      5    5.172 ms      4.09x |  298.044 ms      3.96x |
      6    4.809 ms      4.40x |  262.241 ms      4.51x |
      7    5.011 ms      4.22x |  214.813 ms      5.50x |
      8    4.714 ms      4.49x |  187.067 ms      6.32x |
      9    4.400 ms      4.80x |  177.751 ms      6.65x |
     10    4.391 ms      4.81x |  160.330 ms      7.37x |
     11    4.198 ms      5.04x |  158.244 ms      7.47x |
     12    4.143 ms      5.10x |  147.618 ms      8.00x |
     13    4.060 ms      5.21x |  152.322 ms      7.76x |
     14    4.063 ms      5.20x |  144.492 ms      8.18x |
     15    4.078 ms      5.18x |  135.568 ms      8.71x |
     16    4.027 ms      5.25x |  139.883 ms      8.45x |
     17    3.982 ms      5.31x |  124.331 ms      9.50x |
     18    3.983 ms      5.31x |  124.979 ms      9.45x |
     19    3.990 ms      5.30x |  124.930 ms      9.46x |
     20    4.016 ms      5.26x |  117.301 ms     10.07x |
     21    3.993 ms      5.30x |  116.779 ms     10.12x |
     22    4.041 ms      5.23x |  107.413 ms     11.00x |
     23    4.049 ms      5.22x |  107.895 ms     10.95x |
     24    4.064 ms      5.20x |  102.439 ms     11.53x |
     25    4.128 ms      5.12x |   98.065 ms     12.05x |
     26    5.142 ms      4.11x |   96.914 ms     12.19x |
     27    4.494 ms      4.70x |   95.003 ms     12.44x |
     28    4.922 ms      4.30x |   89.200 ms     13.24x |
     29    4.386 ms      4.82x |   92.067 ms     12.83x |
     30    5.571 ms      3.79x |   92.930 ms     12.71x |
     31    4.798 ms      4.41x |   82.024 ms     14.40x |
     32    4.961 ms      4.26x |   95.622 ms     12.36x |
     33    5.144 ms      4.11x |   89.544 ms     13.19x |
     34    5.328 ms      3.97x |   84.976 ms     13.90x |
     35    5.398 ms      3.92x |   86.333 ms     13.68x |
     36    4.845 ms      4.36x |   83.456 ms     14.16x |
     37    5.656 ms      3.74x |   82.850 ms     14.26x |
     38    5.830 ms      3.63x |   83.321 ms     14.18x |
     39    6.929 ms      3.05x |   86.988 ms     13.58x |
     40    5.597 ms      3.78x |   85.253 ms     13.86x |

 ----------------
```

Without `parallel_scan`:
```
Threads          1000          |         10000          |
      1   21.184 ms      1.00x | 1180.022 ms      1.00x |
      2    9.314 ms      2.27x |  610.554 ms      1.93x |
      3    7.021 ms      3.02x |  425.522 ms      2.77x |
      4    5.914 ms      3.58x |  356.583 ms      3.31x |
      5    5.269 ms      4.02x |  307.107 ms      3.84x |
      6    4.837 ms      4.38x |  232.682 ms      5.07x |
      7    4.494 ms      4.71x |  203.451 ms      5.80x |
      8    4.378 ms      4.84x |  198.475 ms      5.95x |
      9    4.489 ms      4.72x |  180.319 ms      6.54x |
     10    4.401 ms      4.81x |  174.229 ms      6.77x |
     11    4.242 ms      4.99x |  171.264 ms      6.89x |
     12    4.166 ms      5.09x |  153.960 ms      7.66x |
     13    4.111 ms      5.15x |  145.475 ms      8.11x |
     14    4.080 ms      5.19x |  141.851 ms      8.32x |
     15    4.062 ms      5.22x |  141.881 ms      8.32x |
     16    4.021 ms      5.27x |  133.542 ms      8.84x |
     17    3.976 ms      5.33x |  129.044 ms      9.14x |
     18    4.024 ms      5.26x |  128.414 ms      9.19x |
     19    4.002 ms      5.29x |  124.770 ms      9.46x |
     20    3.985 ms      5.32x |  128.082 ms      9.21x |
     21    4.039 ms      5.25x |  115.798 ms     10.19x |
     22    4.039 ms      5.24x |  105.702 ms     11.16x |
     23    4.222 ms      5.02x |  109.672 ms     10.76x |
     24    4.062 ms      5.22x |  103.417 ms     11.41x |
     25    4.133 ms      5.13x |  101.364 ms     11.64x |
     26    4.206 ms      5.04x |   95.400 ms     12.37x |
     27    4.260 ms      4.97x |   95.301 ms     12.38x |
     28    4.297 ms      4.93x |   92.711 ms     12.73x |
     29    4.831 ms      4.38x |   81.153 ms     14.54x |
     30    4.420 ms      4.79x |   85.711 ms     13.77x |
     31    5.325 ms      3.98x |   85.350 ms     13.83x |
     32    4.578 ms      4.63x |   86.452 ms     13.65x |
     33    4.706 ms      4.50x |   87.575 ms     13.47x |
     34    5.350 ms      3.96x |   84.085 ms     14.03x |
     35    5.369 ms      3.95x |   87.756 ms     13.45x |
     36    5.907 ms      3.59x |   80.174 ms     14.72x |
     37    5.990 ms      3.54x |   88.833 ms     13.28x |
     38    5.391 ms      3.93x |   87.560 ms     13.48x |
     39    6.929 ms      3.06x |   82.713 ms     14.27x |
     40    6.193 ms      3.42x |   89.340 ms     13.21x |

 ----------------
```

The results are basically identical up to the usual noise. Realistically, the sum is a much smaller part of the calculation than the actual binning and even the rest of the reduction, and in practice the number of bins is generally too small to be worth parallelism, especially in the form of something that actually has to do more work to do the calculation, so it makes sense that you get little to no performance gain. Since there's no meaningful performance loss, removing the extra code to simplify the class is definitely worthwhile.